+++ src/controller.h
@@ -68,7 +68,10 @@
  */
 #undef NETCONF_LOCK_EXTRA_NAMESPACE
 
-#define ACTION_PROCESS "Action process"
+/*! Pyapi or other services process
+ * Note this set also in clispec:s and tests
+ */
+#define SERVICES_PROCESS "Services process"
 
 /*! Controller debug levels
  */