+++ src/controller_rpc.c
@@ -618,7 +618,7 @@ rpc_config_pull(clixon_handle h,
 
     clixon_debug(CLIXON_DBG_CTRL, "");
     /* Initiate new transaction */
-    if ((ret = controller_transaction_new(h, ce->ce_id, ce->ce_username, "pull", &ct, &cberr)) < 0)
+    if ((ret = controller_transaction_new(h, ce->ce_id, clicon_username_get(h), "pull", &ct, &cberr)) < 0)
         goto done;
     if (ret == 0){
         if (netconf_operation_failed(cbret, "application", cbuf_get(cberr))< 0)
@@ -1636,7 +1636,7 @@ rpc_controller_commit(clixon_handle h,
     /* Initiate new transaction.
      * NB: this locks candidate, which always needs to be unlocked, eg by controller_transaction_done
      */
-    if ((ret = controller_transaction_new(h, ce->ce_id, ce->ce_username, cbuf_get(cbtr), &ct, &cberr)) < 0)
+    if ((ret = controller_transaction_new(h, ce->ce_id, clicon_username_get(h), cbuf_get(cbtr), &ct, &cberr)) < 0)
         goto done;
     if (ret == 0){
         if (netconf_operation_failed(cbret, "application", cbuf_get(cberr))< 0)
@@ -2086,7 +2086,7 @@ rpc_connection_change(clixon_handle h,
     pattern = xml_body(xn);
     operation = xml_find_body(xe, "operation");
     cprintf(cbtr, " %s", operation);
-    if ((ret = controller_transaction_new(h, ce->ce_id, ce->ce_username, cbuf_get(cbtr), &ct, &cberr)) < 0)
+    if ((ret = controller_transaction_new(h, ce->ce_id, clicon_username_get(h), cbuf_get(cbtr), &ct, &cberr)) < 0)
         goto done;
     if (ret == 0){
         if (netconf_operation_failed(cbret, "application", cbuf_get(cberr))< 0)
@@ -3190,7 +3190,7 @@ rpc_device_rpc_template_apply(clixon_handle h,
             goto done;
     }
     /* Initiate new transaction */
-    if ((ret = controller_transaction_new(h, ce->ce_id, ce->ce_username, "rpc", &ct, &cberr)) < 0)
+    if ((ret = controller_transaction_new(h, ce->ce_id, clicon_username_get(h), "rpc", &ct, &cberr)) < 0)
         goto done;
     if (ret == 0){
         if (netconf_operation_failed(cbret, "application", cbuf_get(cberr))< 0)